Move AstPragmaChecker transformer to own file
