+++ src/AstPragma.h
@@ -17,16 +17,12 @@
 #pragma once
 
 #include "AstNode.h"
-#include "AstTransformer.h"
 #include <cassert>
 #include <ostream>
 #include <string>
 #include <utility>
-#include <vector>
 
 namespace souffle {
-
-class AstTranslationUnit;
 
 /**
  * @class AstPragma
@@ -68,17 +64,6 @@ protected:
 
     /** Value */
     std::string value;
-};
-
-/** TODO (b-scholz): this should not be here */
-class AstPragmaChecker : public AstTransformer {
-public:
-    std::string getName() const override {
-        return "AstPragmaChecker";
-    }
-
-private:
-    bool transform(AstTranslationUnit&) override;
 };
 
 }  // end of namespace souffle
